Always output instructions on batch scan.
Instructions when batch scanning were only output when using verbose mode. But the verbose mode is really verbose. Users that didn't write the program could have trouble knowing what to do when the first page or the whole document is done when not outputting this help. Since there is no built in help or man page.